move a file to a folder
* Organize the files into folders by type.
