+++ app/src/main/java/io/kong/mypetdiary/activity/MainActivity.java
@@ -1,13 +1,15 @@
-package io.kong.mypetdiary;
+package io.kong.mypetdiary.activity;
 
 import android.graphics.Color;
 import android.os.Bundle;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
-import android.view.WindowManager;
-import android.widget.Button;
 import android.widget.ImageButton;
+
+import io.kong.mypetdiary.R;
+import io.kong.mypetdiary.fragment.HomeFragment;
+import io.kong.mypetdiary.fragment.MyPageFragment;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
