+++ app/src/main/java/io/kong/mypetdiary/activity/SignUpActivity.java
@@ -1,10 +1,14 @@
-package io.kong.mypetdiary;
+package io.kong.mypetdiary.activity;
 
 import android.content.Intent;
+import android.graphics.Color;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.View;
 import android.widget.EditText;
 import android.widget.Spinner;
+
+import io.kong.mypetdiary.R;
 
 public class SignUpActivity extends AppCompatActivity {
 
@@ -14,6 +18,10 @@ public class SignUpActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState){
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_signup);
+
+        View view = getWindow().getDecorView();
+        view.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR);
+        getWindow().setStatusBarColor(Color.parseColor("#f2f2f2"));
 
         Intent intent = getIntent();
         int kakao = intent.getExtras().getInt("kakao");
