+++ app/src/main/java/io/kong/mypetdiary/adapter/HomeListViewAdapter.java
@@ -1,4 +1,4 @@
-package io.kong.mypetdiary;
+package io.kong.mypetdiary.adapter;
 
 import android.content.Context;
 import android.graphics.drawable.Drawable;
@@ -11,12 +11,15 @@ import android.widget.TextView;
 
 import java.util.ArrayList;
 
-public class MyPageListViewAdapter extends BaseAdapter {
+import io.kong.mypetdiary.R;
+import io.kong.mypetdiary.item.HomeListViewItem;
+
+public class HomeListViewAdapter extends BaseAdapter {
     // Adapter에 추가된 데이터를 저장하기 위한 ArrayList
     private ArrayList<HomeListViewItem> listViewItemList = new ArrayList<HomeListViewItem>();
 
     // ListViewAdapter의 생성자
-    public MyPageListViewAdapter() {
+    public HomeListViewAdapter() {
 
     }
 
@@ -51,6 +54,7 @@ public class MyPageListViewAdapter extends BaseAdapter {
         iconImageView.setImageDrawable(listViewItem.getIcon());
         txtTitle.setText(listViewItem.getTitle());
         txtDate.setText(listViewItem.getDesc());
+
 
         return convertView;
     }
