+++ app/src/main/java/io/kong/mypetdiary/item/HomeListViewItem.java
@@ -1,4 +1,4 @@
-package io.kong.mypetdiary;
+package io.kong.mypetdiary.item;
 
 import android.graphics.drawable.Drawable;
