+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.kong.mypetdiary.HomeFragment">
+    tools:context="io.kong.mypetdiary.fragment.HomeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
