+++ app/src/main/res/layout/fragment_my_page.xml
@@ -5,7 +5,7 @@
     android:id="@+id/drawerLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="io.kong.mypetdiary.MyPageFragment">
+    tools:context="io.kong.mypetdiary.fragment.MyPageFragment">
 
 
     <FrameLayout
